SplitUpMultipleHandsToLines() compability fixed
